move eventfilereader to kinesis connector
